add uv to streamlit
