+++ streamlit-uv-Cerebras/README.md
@@ -1,0 +1,9 @@
+## how to run this app
+
+requirements: uv  
+packages `streamlit` and `cerebras_cloud_sdk`
+
+```shell
+uv sync
+uv run streamlit run app/app.py
+```
