+++ streamlit-uv-Cerebras/pyproject.toml
@@ -1,0 +1,10 @@
+[project]
+name = "streamlit-uv-cerebras"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = [
+    "cerebras-cloud-sdk>=1.23.0",
+    "streamlit>=1.42.2",
+]
