--- streamlit/.streamlit/config.toml
@@ -1,2 +1,0 @@
-[server]
-runOnSave = true
